feat(conanfile): integrate Sentry Native SDK and update build scripts
Added support for Sentry Native SDK in the Conan package. This includes creating the Sentry package in the build_dependencies.sh script and updating the conanfile.py to require the appropriate Sentry version based on the operating system. Additionally, a new script for building Sentry on macOS has been introduced.